+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.6
+
+- Fix `RichEdit.defaultFont`.
+
 ## 1.0.5
 
 - `TextFormatted`: added `faceName`.